requirements.txt
asttokens==2.2.1
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
decorator==5.1.1
executing==1.2.0
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
ipython==8.8.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
jmespath==1.0.1
lxml==4.9.1
matplotlib-inline==0.1.6
packaging==23.0
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
Protego==0.2.1
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.14.0
pyOpenSSL==22.1.0
python-dotenv==0.21.0
python-scrapyd-api==2.1.2
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.3
scrapyd==1.4.1
scrapyd-client==1.2.3
service-identity==21.1.0
six==1.16.0
stack-data==0.6.2
tldextract==3.4.0
traitlets==5.8.0
Twisted==22.8.0
typing_extensions==4.4.0
uberegg==0.1.1
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
zope.interface==5.4.0
